Add multi-branch test
