+++ iota-streams-app-channels/examples/src/branching/mod.rs
@@ -1,0 +1,3 @@
+pub mod single_branch;
+pub mod multi_branch;
+pub mod utils;
