+++ iota-streams-app-channels/src/api/tangle/subscriber.rs
@@ -75,11 +75,10 @@ impl Subscriber {
     }
 
     /// Subscribe to a Channel app instance.
-    pub fn subscribe(&mut self, link_to: &Address) -> Result<(Message, Option<Message>)> {
+    pub fn subscribe(&mut self, link_to: &Address) -> Result<Message> {
         //TODO: remove link_to
         let subscribe = self.imp.subscribe(link_to.rel(), MsgInfo::Subscribe).unwrap();
-        let sequenced = self.send_sequence(&link_to);
-        Ok((subscribe, sequenced))
+        Ok(subscribe)
     }
 
     /*
@@ -160,7 +159,7 @@ impl Subscriber {
     pub fn unwrap_sequence<'a>(&mut self, preparsed: Preparsed<'a>) -> Result<Address> {
         let seq_msg = self.imp.handle_sequence(preparsed, MsgInfo::Sequence).unwrap();
         let msg_id = self.gen_msg_id(
-            &Address::from_str(&self.imp.appinst.as_ref().unwrap().appinst.to_string(), &seq_msg.ref_link.to_string()).unwrap(),
+            &Address::new(self.imp.appinst.as_ref().unwrap().appinst.clone(), MsgId::from(seq_msg.ref_link)),
             seq_msg.pubkey,
             seq_msg.seq_num.0,
         );
@@ -190,10 +189,9 @@ impl Subscriber {
             seq_num = seq_state_num;
             if branching {
                 seq_num = 1;
-            } else if retry {
+            } else if retry && seq_num > 2 {
                 seq_num -= 1
             }
-
             let id = self.gen_msg_id(&seq_link, pk.0, seq_num.clone());
             ids.push((pk.0, id, seq_num));
         }
